requirements.txt
google-api-python-client==2.4.0
oauth2client==4.1.3
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
pygtrie==2.4.2